+++ aequilibrae/__init__.py
@@ -1,25 +1,4 @@
-from aequilibrae.log import logger, global_logger
-from aequilibrae.parameters import Parameters
-from aequilibrae.project.data import Matrices
-from aequilibrae.log import Log
-from aequilibrae import matrix
-from aequilibrae import transit
-from aequilibrae import project
-
-from aequilibrae.distribution import Ipf, GravityApplication, GravityCalibration, SyntheticGravityModel
-from aequilibrae.matrix import AequilibraeMatrix
-from aequilibrae import distribution
-from aequilibrae.paths.network_skimming import NetworkSkimming
-from aequilibrae.paths.traffic_class import TrafficClass
-from aequilibrae.paths.vdf import VDF
-from aequilibrae.paths.all_or_nothing import allOrNothing
-from aequilibrae.paths.traffic_assignment import TrafficAssignment
-from aequilibrae.paths.graph import Graph
-from aequilibrae.paths.assignment_paths import AssignmentPaths
-from aequilibrae.project import Project
-from aequilibrae.paths.results import AssignmentResults, SkimResults, PathResults
-
-from aequilibrae import paths
+from aequilibrae.project.project import Project
 
 # When updating the version, one must also update the docs/source/useful_links/version_history.rst file
 version = "1.5.0"
