+++ aequilibrae/paths/graph.py
@@ -648,19 +648,3 @@ class GraphBase(ABC):  # noqa: B024
             value.columns = [col.lower() for col in value.columns]
         super().__setattr__(key, value)
 
-
-class Graph(GraphBase):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-
-class TransitGraph(GraphBase):
-    def __init__(self, config: dict = None, od_node_mapping: pd.DataFrame = None, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self._config = config
-        self.od_node_mapping = od_node_mapping
-        self.mode = "t"
-
-    @property
-    def config(self):
-        return self._config
