+++ aequilibrae/paths/results/__init__.py
@@ -1,17 +1,3 @@
-"""
-=============
-path computation related code
-=============
-
-STILL NEED TO ADD SOME EXPLANATIONS HERE
-
-"""
-
-__author__ = "Pedro Camargo ($Author: Pedro Camargo $)"
-__version__ = "1.0"
-__revision__ = "$Revision: 1 $"
-__date__ = "$Date: 2016-07-02$"
-
 from .assignment_results import AssignmentResults, TransitAssignmentResults
 from .path_results import PathResults
 from .skim_results import SkimResults
