+++ aequilibrae/project/__init__.py
@@ -1,11 +1,1 @@
-from aequilibrae.project.project import Project
-from aequilibrae.project.about import About
-from aequilibrae.project.network.network import Network
-from aequilibrae.project.field_editor import FieldEditor
-from aequilibrae.project.zoning import Zoning
-from aequilibrae.project.zone import Zone
-from aequilibrae.project.data import Matrices
-from aequilibrae.log import Log
-from aequilibrae.project.network.period import Period
-from aequilibrae.project.network.periods import Periods
-from aequilibrae.project.tools.network_simplifier import NetworkSimplifier
+
