--- aequilibrae/project/data/__init__.py
@@ -1,2 +1,0 @@
-from .matrices import Matrices
-from .results import Results
