--- aequilibrae/project/network/__init__.py
@@ -1,11 +1,0 @@
-from .network import Network
-from .mode import Mode
-from .modes import Modes
-from .link_type import LinkType
-from .link_types import LinkTypes
-from .links import Links
-from .link import Link
-from .nodes import Nodes
-from .node import Node
-from .period import Period
-from .periods import Periods
