+++ aequilibrae/project/network/network.py
@@ -40,7 +40,7 @@ class Network(WorkerThread):
 
     def __init__(self, project) -> None:
         WorkerThread.__init__(self, None)
-        from aequilibrae.paths import Graph
+        from aequilibrae.paths.graph import Graph
 
         self.graphs = {}  # type: Dict[Graph]
         self.project = project
@@ -294,7 +294,7 @@ class Network(WorkerThread):
             >>> project.close()
 
         """
-        from aequilibrae.paths import Graph
+        from aequilibrae.paths.graph import Graph
 
         with self.project.db_connection as conn:
             if fields is None:
