--- aequilibrae/project/tools/__init__.py
@@ -1,2 +1,0 @@
-from .network_simplifier import NetworkSimplifier
-from .migration_manager import MigrationManager, Migration, MigrationStatus
