+++ aequilibrae/traffic_assignment/__init__.py
@@ -1,0 +1,3 @@
+from aequilibrae.traffic_assignment.traffic_assignment import TrafficAssignment
+from aequilibrae.traffic_assignment.traffic_class import TrafficClass
+from .graph import Graph
