+++ aequilibrae/traffic_assignment/graph.py
@@ -1,0 +1,6 @@
+from aequilibrae.paths.graph import GraphBase
+
+
+class Graph(GraphBase):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
