+++ aequilibrae/traffic_assignment/traffic_class.py
