--- aequilibrae/transit/__init__.py
@@ -1,2 +1,0 @@
-from .transit import Transit
-from .transit_graph_builder import TransitGraphBuilder
