--- aequilibrae/transit/functions/__init__.py
@@ -1,1 +1,0 @@
-# from .fix_connections import fix_connections_table
