--- aequilibrae/transit/gtfs_writer/__init__.py
@@ -1,7 +1,0 @@
-from .agency_writer import write_agencies
-from .routes_writer import write_routes
-from .shape_writer import write_shapes
-from .stop_times_writer import write_stop_times
-from .stops_writer import write_stops
-from .trips_writer import write_trips
-from .fare_writer import write_fares
