--- aequilibrae/transit/route_system_reader/__init__.py
@@ -1,6 +1,0 @@
-from .agency_reader import read_agencies
-from .pattern_reader import read_patterns
-from .routes_reader import read_routes
-from .stop_reader import read_stops
-from .stop_times_reader import read_stop_times
-from .trips_reader import read_trips
