--- aequilibrae/transit/transit_elements/__init__.py
@@ -1,10 +1,0 @@
-from .agency import Agency
-from .fare import Fare
-from .fare_rule import FareRule
-from .link import Link
-from .pattern import Pattern
-from .route import Route
-from .service import Service
-from .stop import Stop
-from .trip import Trip
-from .mode_correspondence import mode_correspondence
