+++ docs/source/examples/aequilibrae_project/plot_check_logging.py
@@ -21,7 +21,7 @@ from uuid import uuid4
 from tempfile import gettempdir
 from os.path import join
 from aequilibrae.utils.create_example import create_example
-from aequilibrae.paths import TrafficAssignment, TrafficClass
+from aequilibrae.traffic_assignment import TrafficClass, TrafficAssignment
 # sphinx_gallery_thumbnail_path = '../source/_images/logs_image.png'
 
 # %%
