+++ tests/aeq/paths/test_graph.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from aequilibrae.paths import Graph
+from aequilibrae.paths.graph import Graph
 from aequilibrae.paths.results import PathResults
 from aequilibrae.transit import Transit
